backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.136.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
playwright==1.58.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
